Split alarm configurations into their own files
