+++ aws/cloudformation/standalone/gen_ai_curriculum/monitoring/alarms/createAlarms.ts
@@ -7,10 +7,10 @@ import {REGION} from '../constants';
 
 import {
   openaiSafetyHighFailureRateConfiguration,
-  chatCompletionJobExecutionHighFailureRateConfigurationOpenai,
-  chatCompletionJobExecutionHighFailureRateConfigurationSagemaker,
+  chatCompletionJobExecutionHighFailureRateOpenaiConfiguration,
+  chatCompletionJobExecutionHighFailureRateSagemakerConfiguration,
   chatCompletionHighBrowserFailureRateConfiguration,
-} from './alarmConfigurations';
+} from './configurations';
 
 // Initialize the CloudWatch client.
 const cloudwatch = new CloudWatchClient({region: REGION});
@@ -18,8 +18,8 @@ const cloudwatch = new CloudWatchClient({region: REGION});
 // Array of alarm configurations.
 const alarmConfigurations = [
   openaiSafetyHighFailureRateConfiguration,
-  chatCompletionJobExecutionHighFailureRateConfigurationOpenai,
-  chatCompletionJobExecutionHighFailureRateConfigurationSagemaker,
+  chatCompletionJobExecutionHighFailureRateOpenaiConfiguration,
+  chatCompletionJobExecutionHighFailureRateSagemakerConfiguration,
   chatCompletionHighBrowserFailureRateConfiguration,
 ];
